# web/web/src/assets/openapi.yml
openapi: 3.0.3
info:
  title: Werner FYI - API Resources
  description: Data resources compiled for and used by werner.fyi
  version: 1.0.0
servers:
  - url: 'http://api.werner.fyi'
paths:
  /covid-all:
    post:
      tags:
        - "COVID-19"
      summary: Returns all COVID cases or death data for the U.S.
      description: >-
        Provides all COVID cases or death data for the entire U.S. or a U.S. region, state, or county.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CovidAllRequestBody"
      responses:
        '200':
          description: Pandas dataframe records returned in JSON format.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CovidAllCasesResponse"
                  - $ref: "#/components/schemas/CovidAllDeathsResponse"
  /covid-since-date:
    post:
      tags:
        - "COVID-19"
      summary: 'Returns COVID cases/death data by date, split by 2016 vote'
      description: >-
        Provides COVID cases or death data for entire U.S. or a U.S. region by
        date, split by 2016 electoral result
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CovidSinceDateRequestBody"
      responses:
        '200':
          description: Pandas dataframe records returned in JSON format.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CovidSinceDateCasesResponse"
                  - $ref: "#/components/schemas/CovidSinceDateDeathsResponse"
  /covid-since-first:
    post:
      tags:
        - "COVID-19"
      summary: 'Returns COVID cases/death data by day of first case, split by 2016 vote'
      description: >-
        Provides COVID cases or death data for entire U.S. or a U.S. region by
        day of first case, split by 2016 electoral result
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CovidSinceFirstRequestBody"
      responses:
        '200':
          description: Pandas dataframe records returned in JSON format.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CovidSinceFirstCasesResponse"
                  - $ref: "#/components/schemas/CovidSinceFirstDeathsResponse"
...
components:
  schemas:
    CovidAllRequestBody:
      type: object
      required:
        - category
        - region
        - state
        - county
      properties:
        category:
          $ref: "#/components/schemas/category"
        region:
          $ref: "#/components/schemas/region"
        state:
          $ref: "#/components/schemas/state"
        county:
          $ref: "#/components/schemas/county"
    CovidSinceDateRequestBody:
      type: object
      required:
        - category
        - region
      properties:
        category:
          $ref: "#/components/schemas/category"
        region:
          $ref: "#/components/schemas/region"
    CovidSinceFirstRequestBody:
      type: object
      required:
        - category
        - region
      properties:
        category:
          $ref: "#/components/schemas/category"
        region:
          $ref: "#/components/schemas/region"
    CovidAllCasesResponse:
      type: object
      items:
        type: object
        required:
          - Date
          - Total Cases
          - New Cases
          - Six Day Rolling Average New Cases
        properties:
          Date:
            $ref: "#/components/schemas/Date"
          Total Cases:
            $ref: "#/components/schemas/TotalCases"
          New Cases:
            $ref: "#/components/schemas/NewCases"
          Six Day Rolling Average New Cases:
            $ref: "#/components/schemas/SixDayRollingAverageNewCases"
    CovidAllDeathsResponse:
      type: object
      items:
        type: object
        required:
          - Date
          - Total Deaths
          - New Deaths
          - Six Day Rolling Average New Deaths
        properties:
          Date:
            $ref: "#/components/schemas/Date"
          Total Deaths:
            $ref: "#/components/schemas/TotalDeaths"
          New Deaths:
            $ref: "#/components/schemas/NewDeaths"
          Six Day Rolling Average New Deaths:
            $ref: "#/components/schemas/SixDayRollingAverageNewDeaths"
    CovidSinceDateCasesResponse:
      type: object
      items:
        type: object
        required:
          - State Winner
          - Date
          - Total Cases
          - New Cases
        properties:
          State Winner:
            $ref: "#/components/schemas/StateWinner"
          Date:
            $ref: "#/components/schemas/Date"
          Total Cases:
            $ref: "#/components/schemas/TotalCases"
          New Cases:
            $ref: "#/components/schemas/NewCases"
    CovidSinceDateDeathsResponse:
      type: object
      items:
        type: object
        required:
          - State Winner
          - Date
          - Total Deaths
          - New Deaths
        properties:
          State Winner:
            $ref: "#/components/schemas/StateWinner"
          Date:
            $ref: "#/components/schemas/Date"
          Total Deaths:
            $ref: "#/components/schemas/TotalDeaths"
          New Deaths:
            $ref: "#/components/schemas/NewDeaths"
    CovidSinceFirstCasesResponse:
      type: object
      items:
        type: object
        required:
          - State Winner
          - Days Since First Case
          - Total Cases
          - New Cases
        properties:
          State Winner:
            $ref: "#/components/schemas/StateWinner"
          Days Since First Case:
            $ref: "#/components/schemas/DaysSinceFirstCase"
          Total Cases:
            $ref: "#/components/schemas/TotalCases"
          New Cases:
            $ref: "#/components/schemas/NewCases"
    CovidSinceFirstDeathsResponse:
      type: object
      items:
        type: object
        required:
          - State Winner
          - Days Since First Death
          - Total Deaths
          - New Deaths
        properties:
          State Winner:
            $ref: "#/components/schemas/StateWinner"
          Days Since First Case:
            $ref: "#/components/schemas/DaysSinceFirstCase"
          Total Deaths:
            $ref: "#/components/schemas/TotalCases"
          New Deaths:
            $ref: "#/components/schemas/NewCases"
...
    StateWinner:
      type: string
      enum:
        - Clinton won
        - Trump won
    Date:
      type: string
      example: "12-31-2020"
    DaysSinceFirstCase:
      type: integer
    NewCases:
      type: integer
    NewDeaths:
      type: integer
    TotalCases:
      type: integer
    TotalDeaths:
      type: integer
    SixDayRollingAverageNewCases:
      type: number
    SixDayRollingAverageNewDeaths:
      type: number
    category:
      type: string
      default: "Cases"
      enum:
        - Cases
        - Deaths
    region:
      type: string
      default: "All"
      enum:
        - All
        - Midwest
        - Northeast
        - South
        - West
    state:
      type: string
      default: "All"
      enum:
        - All
        - Alaska
        - Alabama
        - Arkansas
        - Arizona
        - California
        - Colorado
        - Connecticut
        - District of Columbia
        - Delaware
        - Florida
        - Georgia
        - Hawaii
        - Iowa
        - Idaho
        - Illinois
        - Indiana
        - Kansas
        - Kentucky
        - Louisiana
        - Massachusetts
        - Maryland
        - Maine
        - Michigan
        - Minnesota
        - Missouri
        - Mississippi
        - Montana
        - North Carolina
        - North Dakota
        - Nebraska
        - New Hampshire
        - New Jersey
        - New Mexico
        - Nevada
        - New York
        - Ohio
        - Oklahoma
        - Oregon
        - Pennsylvania
        - Rhode Island
        - South Carolina
        - South Dakota
        - Tennessee
        - Texas
        - Utah
        - Virginia
        - Vermont
        - Washington
        - Wisconsin
        - West Virginia
        - Wyoming
...